[refactor]: TokenDto 리팩토링
