+++ src/main/java/com/community/amkorea/auth/service/AuthService.java
@@ -2,7 +2,7 @@ package com.community.amkorea.auth.service;
 
 import com.community.amkorea.auth.dto.SignInDto;
 import com.community.amkorea.auth.dto.SignUpDto;
-import com.community.amkorea.global.Util.Jwt.TokenDto;
+import com.community.amkorea.global.Util.Jwt.dto.TokenDto;
 
 public interface AuthService {
