+++ src/main/java/com/community/amkorea/global/Util/Jwt/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package com.community.amkorea.global.Util.Jwt;
+package com.community.amkorea.global.Util.Jwt.dto;
 
 import lombok.Builder;
 import lombok.Data;
